Updated compose to remove alpha2.  Networks area expanded to include example network concepts!
Added my actual docker-external network for all to see, learn, and query!  
This is a `docker-external` network is specific to my home network and permits containers to appear as normal hosts on the network.  I love it.  However, I'm actually having some trouble with the macvlan for this project and haven't spent adequate time diagnosing, but I wonder, "Overlay Network. And Nevermore."