Update nginx.conf
Conversation with Gavin this afternoon was good and later I had a "duh" moment.  
He correctly pointed out here "why are you adjusting any of this stuff and not just simply passing it?"  Yes.  Yes indeed!  So scrap all of that junk per his suggestion.  Now, here is where he'll hit his head: in theory /adore-djatoka and /adore-djatoka/resolver are ACCEPTABLE values becomes the PROXY should ALWAYS resolve that to a valid $server_name/path (that's nginx's variable for servername).  
So for testing I've used both https://belle.benjaminrosner.com/adore-djatoka and just /adore-djatoka - from external IPs and both work (this is in paged content) and the same holds true for the OSD setting.  YMMV but this is odd, right @g7morris?

Also wanted to flesh out the upstream definitions for clarity and like, literally saving myself a million headaches.